feat(roles): JS/WASM role interfaces; portable bagging/random-subspace draws
- JS: NativeEstimator + typed Regressor/Transformer interfaces, classes
  generated from the manifest (same names as n4m.roles), N4ME import /
  export; the shared fixture replays and refits identically in WASM.
- Fix: bagging and random-subspace PLS drew samples with
  std::uniform_int_distribution / std::shuffle, which are implementation
  defined (libstdc++ vs libc++/Emscripten/macOS). They now use explicit
  rejection-bounded draws and Fisher-Yates on mt19937_64, so fits are
  identical on every platform; fixtures regenerated.
- JS Status enum realigned with n4m_status_t codes.
- R: WeightedPLS formula fits now take the native affine path.

+++ cpp/src/core/estimator/n4me.cpp
@@ -216,7 +216,7 @@ n4m_status_t decode_state(n4m_context_t* ctx, const unsigned char* bytes, std::s
         }
         n4m_status_t st = N4M_OK;
         if (is_int_type(type)) {
-            std::vector<std::int64_t> v(count);
+            std::vector<std::int64_t> v(static_cast<std::size_t>(count));
             for (auto& x : v) {
                 std::uint64_t u = 0;
                 r.u64(u);
@@ -225,7 +225,7 @@ n4m_status_t decode_state(n4m_context_t* ctx, const unsigned char* bytes, std::s
             st = params.set_ints(name.c_str(), static_cast<n4m_method_param_type_t>(type), v.data(),
                                  static_cast<std::int64_t>(count));
         } else {
-            std::vector<double> v(count);
+            std::vector<double> v(static_cast<std::size_t>(count));
             for (auto& x : v) r.f64(x);
             st = params.set_doubles(name.c_str(), static_cast<n4m_method_param_type_t>(type), v.data(),
                                     static_cast<std::int64_t>(count));
@@ -245,7 +245,8 @@ n4m_status_t decode_state(n4m_context_t* ctx, const unsigned char* bytes, std::s
     for (StateBlock& b : blocks) {
         std::uint64_t len = 0;
         const unsigned char* data = nullptr;
-        if (!r.u32(b.tag) || !r.u64(len) || len > r.remaining() || !r.take(len, data)) {
+        if (!r.u32(b.tag) || !r.u64(len) || len > r.remaining() ||
+            !r.take(static_cast<std::size_t>(len), data)) {
             return corrupt(ctx, "truncated N4ME state block");
         }
         b.bytes.assign(data, data + len);
